refactor: Extract state classes.
