+++ app/src/main/java/no/usn/mob3000/domain/viewmodel/auth/LoginViewModel.kt
@@ -8,8 +8,9 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import no.usn.mob3000.domain.model.AuthError
-import no.usn.mob3000.domain.model.User
+import no.usn.mob3000.domain.model.auth.error.AuthError
+import no.usn.mob3000.domain.model.auth.User
+import no.usn.mob3000.domain.model.auth.state.LoginState
 import no.usn.mob3000.domain.usecase.auth.LoginUseCase
 
 /**
@@ -80,37 +81,4 @@ class LoginViewModel(
     fun resetState() {
         _loginState.value = LoginState.Idle
     }
-}
-
-/**
- * Sealed class representing the possible states of the login process.
- *
- * @author frigvid
- * @created 2024-10-22
- */
-sealed class LoginState {
-    /**
-     * Initial state before any login attempt.
-     */
-    data object Idle : LoginState()
-
-    /**
-     * State while login is in progress.
-     */
-    data object Loading : LoginState()
-
-    /**
-     * State when login succeeds, containing the user data.
-     *
-     * @param user The [User] object.
-     */
-    data class Success(val user: User) : LoginState()
-
-    /**
-     * State when login fails, containing the specific error type.
-     *
-     * @param error The returned [AuthError].
-     * @see AuthError
-     */
-    data class Error(val error: AuthError) : LoginState()
 }
