+++ app/src/main/java/no/usn/mob3000/domain/viewmodel/auth/LogoutViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import no.usn.mob3000.domain.model.auth.state.LogoutState
 import no.usn.mob3000.domain.usecase.auth.LogoutUseCase
 
 /**
@@ -52,38 +53,4 @@ class LogoutViewModel(
     fun resetState() {
         _logoutState.value = LogoutState.Idle
     }
-}
-
-/**
- * Sealed class representing the possible states of the logout process.
- *
- * @author frigvid
- * @created 2024-10-22
- */
-sealed class LogoutState {
-    /**
-     * Initial state before any logout attempt.
-     */
-    data object Idle : LogoutState()
-
-    /**
-     * State while logout is in progress.
-     */
-    data object Loading : LogoutState()
-
-    /**
-     * State when logout succeeds.
-     */
-    data object Success : LogoutState()
-
-    /**
-     * State when logout fails, containing the specific error type.
-     *
-     * Unlike the other authentication error flows, this doesn't really
-     * need its own error type. It throws a `RestException`, but given
-     * this is for logging out, simply returning that should be enough.
-     *
-     * @param error The returned [Exception].
-     */
-    data class Error(val error: Exception) : LogoutState()
 }
