+++ app/src/main/java/no/usn/mob3000/domain/viewmodel/auth/RegistrationViewModel.kt
@@ -5,7 +5,8 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import no.usn.mob3000.domain.model.auth.RegistrationError
+import no.usn.mob3000.domain.model.auth.error.RegistrationError
+import no.usn.mob3000.domain.model.auth.state.RegistrationState
 import no.usn.mob3000.domain.usecase.auth.RegisterUseCase
 
 /**
@@ -71,35 +72,4 @@ class RegistrationViewModel(
     fun resetState() {
         _registrationState.value = RegistrationState.Idle
     }
-}
-
-/**
- * Sealed class representing the possible states of the registration process.
- *
- * @author frigvid
- * @created 2024-11-03
- */
-sealed class RegistrationState {
-    /**
-     * Initial state before any attempt.
-     */
-    data object Idle : RegistrationState()
-
-    /**
-     * State while in progress.
-     */
-    data object Loading : RegistrationState()
-
-    /**
-     * State when success.
-     */
-    data object Success : RegistrationState()
-
-    /**
-     * State when login fails, containing the specific error type.
-     *
-     * @param error The returned [RegistrationError].
-     * @see RegistrationError
-     */
-    data class Error(val error: RegistrationError) : RegistrationState()
 }
